Add legacy option to plugin-proposal-decorators.
