+++ packages/babel-plugin-proposal-decorators/src/transformer.js
@@ -1,0 +1,3 @@
+// Not implemented yet
+
+export default {};
