+++ packages/babel-plugin-proposal-decorators/test/fixtures/decl-to-expression/options.json
@@ -1,4 +1,4 @@
 {
   "presets": [],
-  "plugins": ["proposal-decorators"]
+  "plugins": [["proposal-decorators", { "legacy": true }]]
 }
