+++ packages/babel-plugin-syntax-decorators/src/index.js
@@ -1,11 +1,16 @@
 import { declare } from "@babel/helper-plugin-utils";
 
-export default declare(api => {
+export default declare((api, options) => {
   api.assertVersion(7);
+
+  const { legacy = false } = options;
+  if (typeof legacy !== "boolean") {
+    throw new Error("'legacy' must be a boolean.");
+  }
 
   return {
     manipulateOptions(opts, parserOpts) {
-      parserOpts.plugins.push("decorators");
+      parserOpts.plugins.push(legacy ? "decorators" : "decorators2");
     },
   };
 });
